airflow/requirements.txt
pandas
numpy
pydantic
requests
psycopg2-binary
sqlalchemy
python-dotenv
astronomer-cosmos>=1.4.3
dbt-core~=1.8
dbt-postgres~=1.8

